Stop overwriting existing files
Existing files were overwitten if a file with same name already existed. Now a
check is made before renaming a file. If file already exists on that path,
rename is aborted.

Fix Issue: #26

+++ trimfilename/utils.py
@@ -13,18 +13,19 @@ def rename_file(oldname, newname, dirPath):
 
     changed_file_name = False
 
-    if ( (newname.find('.') != -1) and
-         (len(oldname)-oldname.rfind('.')) == (len(newname)-newname.rfind('.')) ):
-        old_file_path = os.path.join(dirPath, oldname)
+    if( os.path.exists((os.path.join(dirPath, newname))) is False):
+        if ( (newname.find('.') != -1) and
+             (len(oldname)-oldname.rfind('.')) == (len(newname)-newname.rfind('.')) ):
+            old_file_path = os.path.join(dirPath, oldname)
 
-        # Does not rename file if it begins with '.' or the whole file name gets
-        # deleted after rename and also if there is no change in filename
-        if newname and newname[0] != '.' and oldname != newname:
-            os.rename(old_file_path, os.path.join(dirPath, newname))
-            # print('Successfully renamed '+pathNname+' to'
-            #       ' '+ newname)
+            # Does not rename file if it begins with '.' or the whole file name gets
+            # deleted after rename and also if there is no change in filename
+            if newname and newname[0] != '.' and oldname != newname:
+                os.rename(old_file_path, os.path.join(dirPath, newname))
+                # print('Successfully renamed '+pathNname+' to'
+                #       ' '+ newname)
 
-            changed_file_name = True
-    # else log for extension changed
+                changed_file_name = True
+        # else log for extension changed
 
     return changed_file_name
